spec(configuration): seperate the configs to seperate files
