+++ spec/Configuration/Config3.cpp
@@ -1,0 +1,20 @@
+#include <cstdlib>
+
+#include "constants.hpp"
+#include "Configuration.hpp"
+#include "Configuration/Directive/Socket.hpp"
+#include "Configuration/Directive/Block/Main.hpp"
+#include "Configuration/Directive/Block/Http.hpp"
+#include "Configuration/Directive/Block/Server.hpp"
+#include "Configuration/Directive/Block/Location.hpp"
+#include "Configuration/Directive/Block/Events.hpp"
+#include "Configuration/Directive/Simple.hpp"
+#include "Configuration/Directive/Simple/Listen.hpp"
+#include "Configuration/Directive/Simple/ServerName.hpp"
+#include "Configuration/Directive/Simple/Cgi.hpp"
+#include "Configuration/Directive/Simple/Return.hpp"
+
+void  Config3(directive::MainBlock&)
+{
+
+}
