+++ spec/Configuration/ConfigurationT.cpp
@@ -41,6 +41,10 @@ TEST(TestConfiguration0, query_empty)
   ASSERT_TRUE(result.is_empty());
 }
 
+////////////////////////////////////////////////
+////////////   Tests for Config 1   ////////////
+////////////////////////////////////////////////
+
 TEST_F(TestConfiguration1, query_wrong_socket)
 {
   const ConfigurationQueryResult result = config_.query(2, "hi.com", "/omg/what.html");
@@ -141,160 +145,9 @@ TEST_F(TestConfiguration1, query_correct_location_property2)
   ASSERT_EQ(result.location_property->error_log, "/var/logs/error.log");
 }
 
-/**
- * events {
- *  worker_connections 512;
- * }
- * 
- * http {
- *   error_log /var/logs/error.log;
- *   access_log /var/logs/access.log;
- *   server {
- *    listen 80;
- *    server_name hi.com wtf.fr;
- *    location / {
- *      root /var/www/html;
- *      index index.html index.htm default.html;
- * 
- *      location /omg/ {
- *        root /var/www/omg;
- *        cgi js /opt/homebrew/bin/node;
- *      }
- *    }
- *  
- *    autoindex on;
- *    location /images/ {
- *      root /var/www/images;
- *      cgi php /usr/local/bin/php4.3;
- *    }
- * 
- *    location /develop {
- *      return 301 http://www.develop.com;
- *    }
- * }
-*/
-void  Config1(directive::MainBlock& config)
-{
-  {
-    directive::EventsBlock*  events = new directive::EventsBlock();
-    config.add_directive(events);
-    directive::WorkerConnections*  worker_connections = new directive::WorkerConnections(1024);
-    worker_connections->set(512);
-    events->add_directive(worker_connections);
-  }
-  directive::HttpBlock*    http = new directive::HttpBlock();
-  config.add_directive(http);
-
-  {
-    directive::ErrorLog*  error_log = new directive::ErrorLog();
-    error_log->set("/var/logs/error.log");
-    http->add_directive(error_log);
-  }
-
-  {
-    directive::AccessLog*  access_log = new directive::AccessLog();
-    access_log->set("/var/logs/access.log");
-    http->add_directive(access_log);
-  }
-
-  {
-    directive::ServerBlock*  server = new directive::ServerBlock();
-    http->add_directive(server);
-
-    {
-      directive::Listen*  listen = new directive::Listen();
-      listen->add(directive::Socket("80"));
-      server->add_directive(listen);
-    }
-
-    {
-      directive::ServerName*  server_name = new directive::ServerName();
-      server_name->add("hi.com");
-      server_name->add("wtf.fr");
-      server->add_directive(server_name);
-    }
-
-    {
-      directive::LocationBlock*  location = new directive::LocationBlock();
-      location->set("/");
-      server->add_directive(location);
-
-      {
-        directive::Root*  root = new directive::Root();
-        root->set("/var/www/html");
-        location->add_directive(root);
-      }
-
-      {
-        // Construction of multiple indexes on the same line
-        // has to be done in reverse order.
-        directive::Index*  index = new directive::Index();
-        index->set("default.html");
-        location->add_directive(index);
-        index = new directive::Index();
-        index->set("index.htm");
-        location->add_directive(index);
-        index = new directive::Index();
-        index->set("index.html");
-        location->add_directive(index);
-      }
-        
-      {
-        directive::LocationBlock*  location2 = new directive::LocationBlock();
-        location2->set("/omg/");
-        location->add_directive(location2);
-
-        {
-          directive::Root*  root = new directive::Root();
-          root->set("/var/www/omg");
-          location2->add_directive(root);
-        }
-
-        {
-          directive::Cgi*  cgi = new directive::Cgi();
-          cgi->set("js", "/opt/homebrew/bin/node");
-          location2->add_directive(cgi);
-        }
-      }
-    }
-
-    {
-      directive::Autoindex*  autoindex = new directive::Autoindex();
-      autoindex->set(true);
-      server->add_directive(autoindex);
-    }
-
-    {
-      directive::LocationBlock*  location = new directive::LocationBlock();
-      location->set("/images/");
-      server->add_directive(location);
-
-      {
-        directive::Root*  root = new directive::Root();
-        root->set("/var/www/images");
-        location->add_directive(root);
-      }
-
-      {
-        directive::Cgi*  cgi = new directive::Cgi();
-        cgi->set("php", "/usr/local/bin/php4.3");
-        location->add_directive(cgi);
-      }
-    }
-
-    {
-      directive::LocationBlock*  location = new directive::LocationBlock();
-      location->set("/develop");
-      server->add_directive(location);
-
-      {
-        directive::Return*  return_ = new directive::Return();
-        return_->set("http://www.develop.com", 301);
-        location->add_directive(return_);
-      }
-    }
-  }
-}
+///////////////////////////////////////////////
+////////////   Test for Config 2   ////////////
+///////////////////////////////////////////////
 
 TEST_F(TestConfiguration2, query_wrong_socket)
 {
@@ -305,135 +158,4 @@ TEST_F(TestConfiguration2, query_wrong_socket)
 TEST_F(TestConfiguration2, query_server_block)
 {
   ASSERT_EQ(config_.query_server_block(3, "hi.com"), &*main_block_->http().servers().first->second);
-}
-
-/**
- * http {
- *   server {
- *     listen 198.13.20.5:443;
- *     listen [;;]:443;
- *     server_name hi.com;
- *     root /var/www/original;
- * 
- *     location /images {
- *       root /var/example;
- *     }
- *   }
- * 
- *   server {
- *     root what/hello/;
- *     cgi php /usr/local/bin/php4.3;
- *     cgi js /opt/homebrew/bin/node;
- *     cgi php /usr/local/bin/php4.8;
- *   }
- * 
- *   server {
- *      listen 127.0.0.1;
- *      server_name wtf.fr;
- *      autoindex on;
- *   }}
- * }
-*/
-void  Config2(directive::MainBlock& main)
-{
-  directive::HttpBlock*    http = new directive::HttpBlock();
-  main.add_directive(http);
-
-  // server 1
-  {
-    directive::ServerBlock*  server = new directive::ServerBlock();
-    http->add_directive(server);
-
-    {
-      directive::Listen*  listen = new directive::Listen();
-      listen->add(directive::Socket("198.13.20.5", "443"));
-      server->add_directive(listen);
-    }
-    {
-      directive::Listen*  listen = new directive::Listen();
-      listen->add(directive::Socket("::", "443"));
-      server->add_directive(listen);
-    }
-    {
-      directive::ServerName*  server_name = new directive::ServerName();
-      server_name->add("hi.com");
-      server->add_directive(server_name);
-    }
-    {
-      directive::Root*  root = new directive::Root();
-      root->set("/var/www/original");
-      server->add_directive(root);
-    }
-
-    {
-      directive::LocationBlock*  location = new directive::LocationBlock();
-      location->set("/images");
-      server->add_directive(location);
-
-      {
-        directive::Root*  root = new directive::Root();
-        root->set("/var/example");
-        location->add_directive(root);
-      }
-    }
-  }
-
-  // server 2
-  {
-    directive::ServerBlock*  server = new directive::ServerBlock();
-    http->add_directive(server);
-
-    {
-      directive::Root*  root = new directive::Root();
-      root->set("what/hello/");
-      server->add_directive(root);
-    }
-
-    {
-      directive::Cgi*  cgi = new directive::Cgi();
-      cgi->set("php", "/usr/local/bin/php4.3");
-      server->add_directive(cgi);
-    }
-
-    {
-      directive::Cgi*  cgi = new directive::Cgi();
-      cgi->set("js", "/opt/homebrew/bin/node");
-      server->add_directive(cgi);
-    }
-
-    {
-      directive::Cgi*  cgi = new directive::Cgi();
-      cgi->set("php", "/usr/local/bin/php4.8");
-      server->add_directive(cgi);
-    }
-  }
-
-  // server 3
-  {
-    directive::ServerBlock*  server = new directive::ServerBlock();
-    http->add_directive(server);
-
-    {
-      directive::Listen*  listen = new directive::Listen();
-      listen->add(directive::Socket("127.0.0.1"));
-      server->add_directive(listen);
-    }
-
-    {
-      directive::ServerName*  server_name = new directive::ServerName();
-      server_name->add("wtf.fr");
-      server->add_directive(server_name);
-    }
-
-    {
-      directive::Autoindex*  autoindex = new directive::Autoindex();
-      autoindex->set(true);
-      server->add_directive(autoindex);
-    }
-  }
-}
-
-void  Config3(directive::MainBlock&)
-{
-
 }
